Fix incorrect routing implementation
