+++ src/Router.js
@@ -1,21 +1,36 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
-  BrowserRouter as Portfolio,
+  HashRouter as Portfolio,
   Route,
   Switch,
  } from 'react-router-dom';
 import Home from './components/home/Home';
 import Projects from './components/projects/Projects';
 import { getRealPathname } from './lib/Helpers';
+import Navbar from './components/home/Navbar';
 
 const Router = () => {
+  const [current, setCurrent] = useState(getRealPathname() === "" ? "home" : getRealPathname());
+  const handleNavbarClick = e => {
+    if (e.key !== "none") {
+      setCurrent(e.key);
+    }
+  };
+
   return (
     <Portfolio>
       <Switch>
-        <div>
-          <Route path={`${getRealPathname()}/`} component={ Home }/>
-          <Route path={`${getRealPathname()}/projects`} component={ Projects }/>
-        </div>
+        <React.Fragment>
+          <Navbar
+              current={current}
+              handleNavbarClick={handleNavbarClick}
+            >
+          </Navbar>
+          <div>
+            <Route path="/" component={ Home }/>
+            <Route path="/projects" component={ Projects }/>
+          </div>
+        </React.Fragment>
       </Switch>
     </Portfolio>
   )
