+++ src/lib/Helpers.js
@@ -3,3 +3,7 @@ export const getRealPathname = () => {
   const realPathname = pathname.includes("portfolio") ? pathname.replace("/portfolio/", "") : pathname.replace(/\//g, "");
   return realPathname;
 }
+
+export const getRealRoute = () => {
+  return window.location.pathname.replace(/\//g, "");
+}
